+++ packages/plugin-packages/mashroom-security/src/services/MashroomSecurityService.ts
@@ -201,14 +201,15 @@ export default class MashroomSecurityService implements MashroomSecurityServiceT
         const securityProvider = this._getSecurityProvider(logger);
         if (securityProvider) {
             try {
-                // If this is a html request, create a new session to prevent session fixation attacks
-                // See https://owasp.org/www-community/attacks/Session_fixation
-                if (isHtmlRequest(request)) {
+                // Only authenticate regular requests if user interaction is required
+                if (isHtmlRequest(request) || await securityProvider.canAuthenticateWithoutUserInteraction(request)) {
+                    // Create a new session to prevent session fixation attacks
+                    // See https://owasp.org/www-community/attacks/Session_fixation
                     await this._createNewSessionIfAny(request);
+                    const authenticationHints = this._getAuthenticationHints(request);
+                    this._removeAuthenticationHintsFromUrl(request, authenticationHints);
+                    return await securityProvider.authenticate(request, response, authenticationHints);
                 }
-                const authenticationHints = this._getAuthenticationHints(request);
-                this._removeAuthenticationHintsFromUrl(request, authenticationHints);
-                return await securityProvider.authenticate(request, response, authenticationHints);
             } catch (e) {
                 logger.error('Security provider returned error: ', e);
             }